feat: 로딩, 에러 화면 적용
