+++ src/components/auth/KakaoRedirect.jsx
@@ -1,12 +1,16 @@
-import { useCallback, useEffect } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import useAuth from "@/contexts/useAuth";
 import socialSignIn from "@/api/auth/socialSignIn";
+import Loading from "@/components/common/Loading";
+import Error from "@/components/common/Error";
 
 const KakaoRedirect = () => {
   const location = useLocation();
   const navigate = useNavigate();
   const { login } = useAuth();
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState(null);
 
   const fetchKakaoAccessToken = useCallback(
     async (code) => {
@@ -20,6 +24,9 @@ const KakaoRedirect = () => {
         }
       } catch (error) {
         console.error("카카오 로그인 중 오류 발생:", error.message);
+        setError(error);
+      } finally {
+        setIsLoading(false);
       }
     },
     [navigate, login]
@@ -32,6 +39,9 @@ const KakaoRedirect = () => {
       fetchKakaoAccessToken(code);
     }
   }, [location, fetchKakaoAccessToken]);
+
+  if (isLoading) return <Loading />;
+  if (error) return <Error />;
 };
 
 export default KakaoRedirect;
