+++ src/components/home/Rank.jsx
@@ -1,44 +1,17 @@
-import getStocksRank from "@/api/stocks/getStocksRank";
-import { useCallback, useEffect, useState } from "react";
+import PropTypes from "prop-types";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
-const Rank = () => {
+const Rank = ({ rank, rankPage, type, types, time, setRankPage, setType }) => {
   const navigate = useNavigate();
-  const [rank, setRank] = useState([]);
-  const [type, setType] = useState("VOLUME");
-  const [time, setTime] = useState("");
-  const [page, setPage] = useState(0);
-
-  const types = [
-    { value: "VOLUME", korean: "거래량" },
-    { value: "RISING", korean: "급상승" },
-    { value: "FALLING", korean: "급하락" },
-  ];
-
-  const fetchRank = useCallback(async () => {
-    try {
-      const response = await getStocksRank({
-        type: type,
-      });
-      setRank(response.data.rank);
-      setTime(response.data.time);
-    } catch (error) {
-      console.error("주식 순위 가져오기 실패: ", error.message);
-    }
-  }, [type]);
-
-  useEffect(() => {
-    fetchRank();
-  }, [fetchRank]);
 
   const handleType = (type) => () => {
     setType(type);
-    setPage(0);
+    setRankPage(0);
   };
 
   const handlePage = (pageNumber) => () => {
-    setPage(pageNumber - 1);
+    setRankPage(pageNumber - 1);
   };
 
   const handleClickStock = (el) => () => {
@@ -79,7 +52,7 @@ const Rank = () => {
           </RankTheadTr>
         </RankThead>
         <RankTableContent>
-          {rank.slice(page * 10, page * 10 + 10).map((el) => {
+          {rank.slice(rankPage * 10, rankPage * 10 + 10).map((el) => {
             const formattedPrice = el.price.toLocaleString();
             const formattedPrdyVrss = el.prdyVrss.toLocaleString();
             const adjustedPrdyCtrt =
@@ -109,7 +82,7 @@ const Rank = () => {
             <PageBtn
               key={index}
               onClick={handlePage(index + 1)}
-              $isActive={page === index}
+              $isActive={rankPage === index}
             >
               {index + 1}
             </PageBtn>
@@ -118,6 +91,16 @@ const Rank = () => {
       </RankTablePageController>
     </RankContainer>
   );
+};
+
+Rank.propTypes = {
+  rank: PropTypes.array.isRequired,
+  rankPage: PropTypes.number.isRequired,
+  type: PropTypes.string.isRequired,
+  types: PropTypes.array.isRequired,
+  time: PropTypes.string.isRequired,
+  setRankPage: PropTypes.func.isRequired,
+  setType: PropTypes.func.isRequired,
 };
 
 export default Rank;
