+++ src/components/stocks/StockChartContainer.jsx
@@ -1,93 +1,8 @@
-import { useState, useEffect } from "react";
 import StockChart from "./StockChart";
-import getStocksChart from "@/api/stocks/getStockChart";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-const StockChartContainer = ({ stock }) => {
-  const [chartData, setChartData] = useState([]);
-  const [period, setPeriod] = useState("DAILY");
-
-  const periods = [
-    { value: "MINUTES", korean: "10분" },
-    { value: "DAILY", korean: "일" },
-    { value: "WEEKLY", korean: "주" },
-    { value: "MONTHLY", korean: "월" },
-    { value: "YEARLY", korean: "년" },
-  ];
-
-  useEffect(() => {
-    const fetchChartData = async () => {
-      try {
-        if (period === "MINUTES") {
-          const now = new Date();
-          const hours = now.getHours();
-          const minutes = now.getMinutes();
-          const isTradingTime =
-            (hours === 9 && minutes >= 10) || (hours > 9 && hours < 16);
-
-          const requests = [
-            getStocksChart({
-              stockCode: stock.stockCode,
-              period: "MINUTES_WEEK",
-            }),
-          ];
-
-          if (isTradingTime) {
-            requests.push(
-              getStocksChart({
-                stockCode: stock.stockCode,
-                period: "MINUTES_TODAY",
-              })
-            );
-          }
-
-          const responses = await Promise.all(requests);
-
-          const transformData = (data) =>
-            data?.map((el) => ({
-              time: el.date,
-              open: el.open,
-              high: el.high,
-              low: el.low,
-              close: el.close,
-              value: el.volume,
-            })) || [];
-
-          const combinedData = responses.flatMap((response) =>
-            transformData(response.data)
-          );
-          combinedData.sort((a, b) => a.time - b.time);
-
-          setChartData(combinedData);
-        } else {
-          const response = await getStocksChart({
-            stockCode: stock.stockCode,
-            period,
-          });
-          const transformedData =
-            response?.data.map((el) => ({
-              time: el.date,
-              open: el.open,
-              high: el.high,
-              low: el.low,
-              close: el.close,
-              value: el.volume,
-            })) || [];
-          setChartData(transformedData);
-        }
-      } catch (error) {
-        console.error("주식 차트 데이터 가져오기 실패:", error.message);
-      }
-    };
-
-    fetchChartData();
-  }, [stock.stockCode, period]);
-
-  const handlePeriod = (period) => () => {
-    setPeriod(period);
-  };
-
+const StockChartContainer = ({ period, periods, handlePeriod, chartData }) => {
   return (
     <StockContainer>
       <ChartContainer>
@@ -113,10 +28,10 @@ const StockChartContainer = ({ stock }) => {
 };
 
 StockChartContainer.propTypes = {
-  stock: PropTypes.shape({
-    stockName: PropTypes.string.isRequired,
-    stockCode: PropTypes.string.isRequired,
-  }),
+  period: PropTypes.string.isRequired,
+  periods: PropTypes.array.isRequired,
+  handlePeriod: PropTypes.func.isRequired,
+  chartData: PropTypes.array.isRequired,
 };
 
 export default StockChartContainer;
